requirements.txt
transformers
bitsandbytes
torch
sentencepiece
faiss-cpu
sentence-transformers